+++ app/src/main/java/com/snister/carnagealpha/features/expense_tracker/domain/repository/LocalRepository.kt
@@ -2,6 +2,6 @@ package com.snister.carnagealpha.features.expense_tracker.domain.repository
 
 interface LocalRepository {
 
-    suspend fun getBalance(): Double
-    suspend fun updateBalance(balance: Double)
+    suspend fun getBalance(): Long
+    suspend fun updateBalance(balance: Long)
 }
