+++ app/src/main/java/com/snister/carnagealpha/features/expense_tracker/presentation/dashboard_overview/DashboardOverviewState.kt
@@ -6,7 +6,7 @@ import java.time.ZonedDateTime
 data class DashboardOverviewState(
     val spendingList: List<SpendingEntity> = emptyList(),
     val datesList: List<ZonedDateTime> = emptyList(),
-    val balance: Double = 0.0,
+    val balance: Long = 0,
     val pickedDate: ZonedDateTime = ZonedDateTime.now(),
     val isDatePickerDropDownMenuVisible: Boolean = false,
 )
